+++ test/testTimeSelect.jsx
@@ -6,6 +6,7 @@ var expect = require('chai')
 .use(require('dirty-chai')).expect;
 
 var Input = require('react-bootstrap').Input;
+
 var ReactIntl = require('react-intl');
 
 var assert = require('assert');
@@ -13,6 +14,8 @@ var sinon = require('sinon');
 var { shallow, mount } = require('enzyme');
 
 var TimeSelect = require('../src/TimeSelect');
+var HourInput = require('../src/HourInput');
+var MinuteInput = require('../src/MinuteInput');
 
 describe('TimeSelect', function() {
   it('is an element', function() {
@@ -31,6 +34,7 @@ describe('TimeSelect', function() {
         name: 'Time',
         label: 'Time',
         locale: 'en-GB',
+        seperateHourMins: false,
         formats: {
           time: {
             short: {
@@ -119,6 +123,45 @@ describe('TimeSelect', function() {
       var timeSelect = shallow(<TimeSelect {...props} id="timeSelect"/>);
       expect(timeSelect.find(Input).prop('id')).to.equal('timeSelect');
     });
+
+    context('seperateHourMins', function() {
+      beforeEach(function() {
+        props = {
+          time: {
+            hours: '10',
+            minutes: '5'
+          },
+          seperateHourMins: true
+        };
+      });
+
+      it('should have a hours input', function() {
+        var timeSelect = mount(<TimeSelect {...props}/>);
+        expect(timeSelect.find(HourInput));
+      });
+
+      it('should have a minutes input', function() {
+        var timeSelect = shallow(<TimeSelect {...props}/>);
+        expect(timeSelect.find(MinuteInput));
+      });
+
+      it('should use current time as value', function() {
+        var timeSelect = shallow(<TimeSelect {...props}/>);
+        expect(timeSelect.find(HourInput).props().value).to.equal(props.time.hours);
+        expect(timeSelect.find(MinuteInput).props().value).to.equal(props.time.minutes);
+      });
+
+      it('fills the hours select box with 24 hours', function() {
+        var timeSelect = mount(<TimeSelect {...props}/>);
+        expect(timeSelect.find(HourInput).props().options).to.have.length(24);
+      });
+
+      it('fills the minutes select box with range from steps prop', function() {
+        props.step = 5;
+        var timeSelect = mount(<TimeSelect {...props}/>);
+        expect(timeSelect.find(MinuteInput).props().options).to.have.length(12);
+      });
+    });
   });
 
   describe('events', function() {
@@ -145,7 +188,7 @@ describe('TimeSelect', function() {
       });
     });
 
-    it('will emit a date up if an option is chosen', function() {
+    it('will emit a time up if an option is chosen', function() {
       var handler = sinon.stub();
       var doc = TestUtils.renderIntoDocument(<TimeSelect onChange={handler} />);
       var node = TestUtils.findRenderedDOMComponentWithTag(doc, 'select');
@@ -156,22 +199,50 @@ describe('TimeSelect', function() {
         }
       });
 
-      sinon.assert.calledWith(handler, new Date(2015, 5, 6, 11, 30, 0, 0));
+      sinon.assert.calledWith(handler, { hours: '11', minutes: '30' });
     });
 
-    it('will use a passed in date for the values of days/months/years', function() {
-      var handler = sinon.stub();
-      var date = new Date(2016, 7, 8);
-      var doc = TestUtils.renderIntoDocument(<TimeSelect onChange={handler} value={date} />);
-      var node = TestUtils.findRenderedDOMComponentWithTag(doc, 'select');
+    context('seperateHourMins', function() {
+      var time;
 
-      TestUtils.Simulate.change(node, {
-        target: {
-          value: '14:30'
-        }
+      beforeEach(function() {
+        time = {
+          hours: '11',
+          minutes: '55'
+        };
       });
 
-      sinon.assert.calledWith(handler, new Date(2016, 7, 8, 14, 30, 0, 0));
+      context('hours', function() {
+        it('will emit hours up if an option is chosen', function() {
+          var handler = sinon.stub();
+          var doc = TestUtils.renderIntoDocument(<TimeSelect seperateHourMins={true} time={time} onChange={handler} />);
+          var node = TestUtils.scryRenderedDOMComponentsWithTag(doc, 'select')[0];
+
+          TestUtils.Simulate.change(node, {
+            target: {
+              value: '14'
+            }
+          });
+
+          sinon.assert.calledWith(handler, { hours: '14', minutes: '55' });
+        });
+      });
+
+      context('minutes', function() {
+        it('will emit minutes up if an option is chosen', function() {
+          var handler = sinon.stub();
+          var doc = TestUtils.renderIntoDocument(<TimeSelect seperateHourMins={true} time={time} onChange={handler} />);
+          var node = TestUtils.scryRenderedDOMComponentsWithTag(doc, 'select')[1];
+
+          TestUtils.Simulate.change(node, {
+            target: {
+              value: '50'
+            }
+          });
+
+          sinon.assert.calledWith(handler, { hours: '11', minutes: '50' });
+        });
+      });
     });
   });
 });