+++ CHANGELOG.md
@@ -1,4 +1,9 @@
 # Change log
+
+### 2.2.0 - 2017-02-24
+- Add a seperateHourMins prop that allows for individual hour and minute inputs. 
+- Add a time prop that passes hour / minute values to hour / minute inputs.
+
 ### 2.1.0 - 2016-12-19
 - Add an id prop that is passed to the input component.